githubAI/requirements.txt
streamlit==1.28.0
torch==2.1
torchvision==0.15.2
# transformers==4.34.0
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
faiss-cpu==1.7.4
pinecone-client==2.2.4
requests==2.31.0
beautifulsoup4==4.12.2
plotly==5.15.0
opencv-python==4.8.0.76
sentence-transformers==2.2.2
#pip install --upgrade transformers huggingface_hub accelerate